AI_Server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
gspread==6.1.2
google-auth==2.34.0
pydantic-settings==2.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
